Split block.py 8/8: extract control-flow blocks → models/control_flow_blocks.py
Part 8/8 (final). ForLoopBlock, WhileLoopBlock, ConditionalBlock + LoopBlockExecutedResult /
compute_conditional_scopes move to control_flow_blocks.py. block.py is now a 495-line facade:
block_base re-export + submodule facade imports + BlockSubclasses union + BlockTypeVar + dispatch
(get_all_blocks, resolve_conditional_merge_edges) + shared helpers only.

Fixes a pydantic forward-ref regression: block.py injects BlockTypeVar into control_flow_blocks and
rebuilds ForLoop/WhileLoop after the discriminated union is complete (they type loop_blocks as
list[BlockTypeVar]); otherwise WorkflowDefinition.model_rebuild() fails. Block-type test patch targets
repointed to their owning modules now that block.py is stripped.
