+++ tests/unit/test_code_block_downloads.py
@@ -19,7 +19,6 @@ import pytest
 from skyvern.forge.sdk.core import skyvern_context
 from skyvern.forge.sdk.core.skyvern_context import SkyvernContext
 from skyvern.forge.sdk.schemas.files import FileInfo
-from skyvern.forge.sdk.workflow.models import block as block_module
 from skyvern.forge.sdk.workflow.models import block_base as block_base_module
 from skyvern.forge.sdk.workflow.models import code_block as code_block_module
 from skyvern.forge.sdk.workflow.models.block import CodeBlock
@@ -103,7 +102,6 @@ def _fake_storage_app(monkeypatch: pytest.MonkeyPatch, *, save, get) -> None:
             execute_code_block_override=AsyncMock(return_value=None),
         ),
     )
-    monkeypatch.setattr(block_module, "app", fake_app)
     monkeypatch.setattr(block_base_module, "app", fake_app)
     monkeypatch.setattr(code_block_module, "app", fake_app)
 
@@ -222,7 +220,7 @@ async def test_code_block_scopes_downloads_to_current_loop_iteration(
             run_id="wr_1",
             loop_internal_state={
                 "downloaded_file_signatures_before_iteration": [
-                    block_module.to_downloaded_file_signature(prev_file),
+                    block_base_module.to_downloaded_file_signature(prev_file),
                 ],
             },
         )
