+++ tests/unit/workflow/test_workflow_trigger_block.py
@@ -13,13 +13,13 @@ from skyvern.forge.sdk.core import skyvern_context
 from skyvern.forge.sdk.core.skyvern_context import SkyvernContext
 from skyvern.forge.sdk.experimentation import providers as providers_module
 from skyvern.forge.sdk.workflow.exceptions import (
+    FailedToFormatJinjaStyleParameter,
     InvalidWorkflowDefinition,
     PayloadTemplateRenderError,
     PayloadTemplateSyntaxError,
 )
 from skyvern.forge.sdk.workflow.models._jinja import _JSON_TYPE_MARKER
 from skyvern.forge.sdk.workflow.models.block import (
-    FailedToFormatJinjaStyleParameter,
     WorkflowTriggerBlock,
     jinja_sandbox_env,
 )
